user2: client test available

+++ server/include/auth.h
@@ -13,7 +13,7 @@ int getUserIDByUsername(MYSQL* pconn, const char* username);
 bool userExist(MYSQL* pconn, const char* username);
 bool userPass(MYSQL* pconn, int uid, const char* cryptpasswd);
 int userInsert(MYSQL* pconn, const char* username, const char* cryptpasswd,
-               int pwdid);
+               long long pwdid);
 int userUpdate(MYSQL* pconn, int uid, const char* fieldname, const char* value);
 
 #endif
